Fix parse_tag_list/parse_swath_timing type hint
When the query tag of parse_tag_list contains subelements of different
kind, the underlying library xmlschema returns a dictionary instead of a
list. This was the case for parse_swath_timing, which contains not only
the burst list but also the number of lines and samples. Instead of
parsing the XML two more times for gathering such information, I
preferred to etend the type hinting.

The drawback: who calls parse_tag_list must make mypy ignore the check
on its return type.